+++ client/commands/local_follow.rs
@@ -12,6 +12,7 @@ pub async fn local_follow(
     pueue_directory: &Path,
     task_id: &Option<usize>,
     err: bool,
+    lines: Option<usize>,
 ) -> Result<()> {
     // The user can specify the id of the task they want to follow
     // If the id isn't specified and there's only a single running task, this task will be used.
@@ -46,7 +47,7 @@ pub async fn local_follow(
         }
     };
 
-    follow_local_task_logs(pueue_directory, task_id, err);
+    follow_local_task_logs(pueue_directory, task_id, err, lines);
 
     Ok(())
 }